requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
gspread>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
unidecode>=1.3.0
streamlit>=1.30.0
streamlit-searchbox>=0.1.14
streamlit-aggrid>=1.0.5
python-telegram-bot>=21.0
sentry-sdk>=2.0.0
